Moved Image resize into operations folder
